coingeco api fixed
