+++ frontend/src/pages/dex.tsx
@@ -1,81 +1,26 @@
 import { Button } from '../components/ui/button';
-import React, {FC, useEffect, useState} from 'react';
+import React, {FC} from 'react';
 import '../styles/App.css';
 import '../styles/tailwind.css';
 import {useNavigate} from "react-router";
 import {ExchangeForm} from "../components/ExchangeForm";
 import {useAccount} from "wagmi";
-import { fetchTransactions } from '../api/transactionsApi';
-import { Transaction } from '../types/types';
 import LoadingWidget from "../components/LoadingWidget";
 import ExchangeProvider from "../utils/ExchangeContext";
 import TransactionTable from "../components/TransactionTable";
+import PriceCard from "../components/PriceCard";
 
 
 const DEX: FC = () => {
 
-    const[ethPrice, setEthPrice] = useState<number>();
 
     const navigate = useNavigate();
     const { address, isConnected } = useAccount();
-    const [transactions, setTransactions] = useState<Transaction[]>([]);
 
     const handleBack = () => {
         console.log('isConnected', isConnected);
         navigate('/');
     }
-
-
-    const fetchPrice = async () => {
-        try {
-            const apiKey = process.env.REACT_APP_COINGECO_API_KEY;
-            if (!apiKey) {
-                throw new Error('API key is missing');
-            }
-
-            const url = 'https://api.coingecko.com/api/v3/simple/price?ids=ethereum&vs_currencies=usd';
-
-            const options = {
-                method: 'GET',
-                headers: {
-                    accept: 'application/json',
-                    'x-cg-pro-api-key': apiKey,
-                },
-            };
-
-            const response = await fetch(url, options);
-            if (!response.ok) {
-                throw new Error(`Error: ${response.statusText}`);
-            }
-
-            const data = await response.json();
-            setEthPrice(data.ethereum.usd);
-            console.log(data.ethereum.usd); // You can now use the fetched data
-
-        } catch (err) {
-            console.error('Error fetching price:', err);
-        }
-    };
-
-    const loadTransactions = async () => {
-        try {
-            const data = await fetchTransactions();
-            setTransactions(data);
-            console.log("Loaded transactions:", data); // Debugging line
-        } catch (error) {
-            console.error("Error loading transactions:", error);
-        }
-    };
-
-    useEffect(() => {
-        // Fetch the price immediately when the component mounts
-        fetchPrice();
-        loadTransactions();
-        //TODO:
-        // const intervalId = setInterval(fetchPrice, 10000); // 10000ms = 10s
-        // Cleanup function to clear the interval when the component unmounts
-        // return () => clearInterval(intervalId);
-    }, []);
 
     return (
         <>
@@ -98,24 +43,13 @@ const DEX: FC = () => {
                             <div className="border border-gray-200 rounded-md p-6 min-h-80 min-w-fit">
                                 <ExchangeForm/>
                             </div>
-                            <div className="border border-gray-200 rounded-md p-6 min-h-80 min-w-fit">
-                                <p>ETH/USD: {ethPrice}</p>
-                            </div>
+                            <PriceCard/>
                         </div>
 
 
 
                         <div className="border border-gray-200 rounded-md p-6 min-h-80 min-w-fit">
-                                {/*<ul>*/}
-                                {/*    {transactions.map((transaction) => (*/}
-                                {/*        <li key={transaction.id} className="p-2 border-b border-gray-200">*/}
-                                {/*            ID: {transaction.id}, Sender: {transaction.transactionSender},*/}
-                                {/*            Amount: {transaction.ethAmount} ETH*/}
-                                {/*        </li>*/}
-                                {/*    ))}*/}
-                                {/*</ul>*/}
                             <TransactionTable/>
-
                         </div>
                     </div>
 
@@ -126,15 +60,6 @@ const DEX: FC = () => {
                     >
                     Back
                         </Button>
-
-                        <Button
-                            variant="destructive"
-                            onClick={loadTransactions}
-                            className="text-lg py-3 px-6"
-                        >
-                            Fetch transactions
-                        </Button>
-
 
                 </div>
             </ExchangeProvider>
